Introduce CertificateManager to be used as a replacement of static methods to manage certificate
This commit introduces  `CertificateManager` which does exactly the same thing than the code
that was in the file Certificate.cs.

Having a class rather than static methods make it easier for testing, mocking, etc.
In order to not introduce a breaking change, static method now calls a singleton instance of
CertificateManager.